Implementada a opção de Apagar os itens

+++ Cria_Lista.py
@@ -4,6 +4,7 @@ o usuario deve ter a possibilidade de
 inserir, apagar e listar valores da sua lista
 nao permita que o programa quebre com erros de indices na lista
 '''
+import os
 
 
 def tempo():
@@ -62,13 +63,15 @@ while True:
         
         #Salvar arquivo txt
         salvar = input('Deseja salvar essa lista? [S][N] ').upper().strip()[0]
+
         if salvar == 'S':
             nome_arquivo = input('Digite o nome do arquivo: ')
             try:
                 arquivo = open(nome_arquivo, 'r+')
             except:
-                arquivo = open(nome_arquivo, 'w+')
-                arquivo.writelines(lista)
+                arquivo = open(nome_arquivo, 'a')
+                for i in lista:
+                    arquivo.write(f'{i}\n')
                 arquivo.close()
         elif salvar == 'N':
             continuar = input('Deseja sair do Programa? ').upper().strip()
@@ -87,20 +90,23 @@ while True:
 
      
                             #Se a opção for para apagar:
-    '''elif opcao == 'A':
+    elif opcao == 'A':
         apgar_toda_lista = input('Deseja apagar todos os itens da lista? ').upper().strip()[0]
 
 #verifica se quer apagar todos os itens da lista
         if apgar_toda_lista == 'S':
         #pergunta o nome do arquivo que deseja apagar
             deleta_nome_arquivo = input('Digite o nome do arquivo que deseja apagar: ')
-            tem_certeza = input(f'Tem certeza que deseja apagar o arquivo {deleta_nome_arquivo}? '
+            tem_certeza = input(f'Tem certeza que deseja apagar o arquivo "{deleta_nome_arquivo.upper()}"? '
                 'Essa ação não podera ser desfeita! ').upper().strip()[0]
             
             #caso deseja apagar o arquivo inteiro
             if tem_certeza == 'S':
-                os.remove(deleta_nome_arquivo)
-                print('Arquivo apagado!')
+                try:
+                    os.remove(deleta_nome_arquivo)
+                    print('Arquivo apagado!')
+                except:
+                    print('Arquivo Não existe')
             #apaga a lista salva na memoria
             elif tem_certeza == 'N':
                 lista.clear()
@@ -126,8 +132,8 @@ while True:
         #caso deseje apagar pelo menos um item da lista.
         elif apgar_toda_lista == 'N':
             #verifica se a lista esta vazia
-            if len(lista)<=0 or lista == ['']:
-                abre_arquivo = input('deseja abrir algum arquivo? ').upper().strip()[0]
+            #if len(lista)<=0 or lista == ['']:
+            abre_arquivo = input('deseja abrir algum arquivo? ').upper().strip()[0]
             
             #caso deseje abrir algum arquivo
             if abre_arquivo == 'S':
@@ -137,29 +143,32 @@ while True:
                     linhas = arquivo.readlines()
                     for i,e in enumerate(linhas):
                         e = e.rstrip('\n') #isso tira a quebra de linha que vem do arquivo
-                        print(i+1,e)
                         lista_comp.append(e)
                     arquivo.close()
                 except:
-                    print(f'{abre_arquivo} não existe!!')
-
-
-
-                        #PAREI AQUI
-
-
-
+                    print(f'{nome_arquivo} não existe!!voltando para o inicio')
+                    tempo()
+                    continue
 
                 qtd_apagar = int(input(f'sua lista possui {len(lista_comp)} itens, quantos deseja apagar: '))
                 for item in range(qtd_apagar):
-                    apagar = input('digite o numero do item que deseja apagar: ')
-                    if apagar.isdigit():
+                    for i,e in enumerate(lista_comp):
+                        print(f'{i+1} - {e}')
+                    apagar = int(input('digite o numero do item que deseja apagar: '))
+                    apagar -= 1
+                    try:
                         lista_comp.pop(apagar)
-                    else:
+                    except:
                         while apagar.isalpha():
                             apagar = input('Digite um numero: ')
-            
-
+                try:
+                    os.remove(nome_arquivo)
+                    arquivo = open(nome_arquivo, 'r+')
+                except:
+                    arquivo = open(nome_arquivo, 'w+')
+                    for i in lista_comp:
+                        arquivo.writelines(f'{i}\n')
+                    arquivo.close()
             #caso nao deseje abrir nenhum arquivo, o sistema irá parar
             elif abre_arquivo == 'N':
                 break
@@ -169,14 +178,14 @@ while True:
                     abre_arquivo = input('Digite N ou S: ')
             
             
-            
-            for i, e in enumerate(lista):
-                print(i,e)              
+            print('sua lista ficaou assim: ')
+            for i, e in enumerate(lista_comp):
+                print(f'{i+1} - {e}')             
 
 
     
 
-    #se a opção for Listar:
+''' #se a opção for Listar:
     elif opcao == 'L':
         for i in lista:
             print(i)
@@ -189,7 +198,7 @@ while True:
 
     #Caso não digite I,A,L ou S
     else:
-        print('Para continuar, voce precisa digitar I, A, L ou S! ', end='')
+        print('Para continuar, voce precisa digitar I, A, L ou S! ', end='')'''
 
 
 #fim do programa
@@ -199,4 +208,4 @@ if len(lista) <= 0:
 else:
     print('Essa é a sua lista: ')
     for i,e in enumerate(lista):
-        print(f'{i+1}° - {e}', end=' ')'''
+        print(f'{i+1}° - {e}', end=' ')
